calculator: add/edit cropparams

+++ calculator/views.py
@@ -1,6 +1,5 @@
 # pyright: reportGeneralTypeIssues=false
-import importlib
-from django.shortcuts import get_object_or_404, render
+from django.shortcuts import get_object_or_404
 from django.template.response import TemplateResponse
 from django.urls.base import reverse_lazy
 from django.views.decorators.http import require_http_methods
@@ -8,6 +7,7 @@ from django.views.generic.detail import DetailView
 from django.views.generic.edit import CreateView, DeleteView
 from calculator.forms import CropModelForm, CropParameterForm, ManagementForm
 from calculator.models import Crop, Management
+from calculator.utils import get_crop_params_list, get_cropmodels
 from register.models import PlantSpecies
 
 """Register CropModels here"""
@@ -26,7 +26,7 @@ class CropDeleteView(DeleteView):
     model = Crop
 
     def get_success_url(self):
-        return reverse_lazy("spaces:area-detail", kwargs={'pk': self.object.area.pk })
+        return reverse_lazy("spaces:area-detail", kwargs={"pk": self.object.area.pk})
 
 
 def crop_update_view(request, pk):
@@ -36,8 +36,9 @@ def crop_update_view(request, pk):
     if request.POST:
         form = CropModelForm(request.POST, instance=crop)
         if form.is_valid():
-            form.save()
-            crop.parameters.all().delete()
+            instance = form.save()
+            if instance.species != crop.species or instance.variety != crop.variety:
+                crop.parameters.all().delete()
 
     # Instantiate the form
     form = CropModelForm(
@@ -46,8 +47,8 @@ def crop_update_view(request, pk):
             "area": crop.area,
             "sowing": crop.sowing,
             "harvest": crop.harvest,
-            "notes": crop.notes
-        }
+            "notes": crop.notes,
+        },
     )
 
     # Set initial value depending on the set crop
@@ -58,67 +59,66 @@ def crop_update_view(request, pk):
     elif crop.has_species():
         form.initial["species"] = crop.species.pk
 
-    # Loop over CROP_MODELS, run the models and store output to the cropmodels list
-    cropmodels = []
-    module = importlib.import_module("calculator.cropmodels")
-    for cm in CROP_MODELS:
-        class_ = getattr(module, cm)
-        m = class_(crop)  # type: ignore
-        if m.can_run():
-            cropmodels.append(m.output())
+    cropparameter_form = CropParameterForm(initial={"crop": crop})
 
     # JSON data to populate the species tom-select
     plantspecies = list(PlantSpecies.objects.all().values("pk", "common_name"))
-    return render(
-        request,
-        "calculator/crop_update.html",
-        {
-            "crop": crop,
-            "form": form,
-            "plantspecies": plantspecies,
-            "cropmodels": cropmodels,
-        },
-    )
+    cropparameter_table = get_crop_params_list(crop)
+    context = {
+        "crop": crop,
+        "form": form,
+        "plantspecies": plantspecies,
+        "cropparameter_table": cropparameter_table,
+        "cropparameter_form": cropparameter_form,
+    }
+    context.update(get_cropmodels(crop, CROP_MODELS))
+    return TemplateResponse(request, "calculator/crop_update.html", context)
 
-def crop_add_parameter_hx(request, pk):
-    form = CropParameterForm()
-    return TemplateResponse(
-        request,
-        'calculator/partials/cropparameter_form.html',
-        { "form": form }
-    )
 
-@require_http_methods(['POST',])
-def crop_update_parameter_hx(request, pk):
+@require_http_methods(
+    [
+        "POST",
+    ]
+)
+def cropparam_update_hx(request, pk):
+    context = {}
     form = CropParameterForm(request.POST)
-    crop = get_object_or_404(Crop, pk = pk)
+    crop = get_object_or_404(Crop, pk=pk)
     if form.is_valid():
-        instance = form.save(commit=False)
-        instance.crop = crop
-        instance.save()
-    
+        form.save()
+    # context["cropparameter_table"] = CropParameterTable(get_crop_params_list(crop))
+    context["cropparameter_table"] = get_crop_params_list(crop)
+    context["cropparameter_form"] = form
+    context["crop"] = crop
+    context.update(get_cropmodels(crop, CROP_MODELS))
+    return TemplateResponse(
+        request, "calculator/partials/cropparams_collapse.html", context
+    )
+
 
 class CropCreateView(CreateView):
     form_class = CropModelForm
     model = Crop
-    template_name = 'calculator/crop_update.html'
+    template_name = "calculator/crop_update.html"
 
     def get_success_url(self):
-        return reverse_lazy("calculator:crop-update", kwargs={'pk': self.object.pk })
-
+        return reverse_lazy("calculator:crop-update", kwargs={"pk": self.object.pk})
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
-        context["plantspecies"] = list(PlantSpecies.objects.all().values("pk", "common_name"))
+        context["plantspecies"] = list(
+            PlantSpecies.objects.all().values("pk", "common_name")
+        )
         return context
 
     def get_initial(self):
         initial = super().get_initial()
         initial = initial.copy()
-        area = self.kwargs['area_id']
+        area = self.kwargs["area_id"]
         if area:
-            initial['area'] = area
+            initial["area"] = area
         return initial
+
 
 class ManagementCreateView(CreateView):
     model = Management
